HTML STRUCTURE
Created the HTML structure for the page with all the dummy content.
Some CSS classes were added in cases where they are obvious. The HTML
structure can change as the styling is added. I split the page into
header and footer and four sections. The four sections are the awesome,
random, testimonial and action section.

On branch html-structure
Changes to be committed:
new file:   css/stylesheet.css
modified:   index.html